services/alert-service/requirements.txt
python-dotenv
psycopg[binary]
fastapi
uvicorn